refactor!: copy some generated OpenAPI classes into main project
Since the idea is to persist the data retrieved from TMDB into a neo4j
database, and use Spring Data Neo4j to do so, I need to annotate the
model classes, and also I don't need all the generated model since not
all of them will be used. So copy what I need into the main project,
rename them for readability and update the code to use the refactored
model.

Keep the OpenAPI generator task in the gradle build in case I lose the
generated code before copying something, but decouple from the build.